Set of lint fixes.
This changes applies a majority of fixes suggested by running make lint.
Apart from application of lint fixes this change adds:

- A python script that is capable of merging a set of clang tidy yaml
  files that can apply the changes. The script allows for setting up
  exclusion and inclusion filters to control how aggressive application
  should be.
- All the TODO tags were fixed manually.
- Fixup of applications that are incorrect.

Test: Run succeeds on mac, win, linux.
Test: Unit test green

+++ android/android-emu/android/skin/generic-event-buffer.cpp
@@ -14,7 +14,7 @@
 
 #include "android/utils/debug.h"
 
-#include <stdio.h>
+#include <cstdio>
 
 void skin_generic_event_buffer_init(SkinGenericEventBuffer* buffer,
                                     SkinGenericEventFlushFunc flush_func) {
@@ -27,7 +27,7 @@ void skin_generic_event_buffer_add(SkinGenericEventBuffer* buffer,
                                    unsigned code,
                                    unsigned value) {
     if (buffer->event_count < MAX_GENERIC_EVENTS) {
-        int count = (int)buffer->event_count++;
+        int count = buffer->event_count++;
         buffer->events[count].type = type;
         buffer->events[count].code = code;
         buffer->events[count].value = value;
